feat(styles): Add styled-components support, global styles, theme, and providers.
Added support for styled-components to the project by creating a styled.d.ts to handle type definitions and a styledComponentsRegistry.tsx for ssr styles support.  Created a globalStyles.ts file to define global styles that can be used across the app. Added a `theme.ts` file to define a base theme with common styles and variables. Created a `providers` directory with an `index.ts` file that exports all of the app's providers, including a `ThemeProvider` that receives the theme prop.