fix: Address "unique key prop" warning in StyledComponentsRegistry
This commit fixes a warning that was being generated in the console:
"Each child in a list should have a unique "key" prop."

To fix this, we've updated the `useServerInsertedHTML()` method to return the child element directly instead of wrapping it in a fragment.

We've also added a temporary `// @ts-ignore` comment to ignore a TypeScript error caused by the `@types/styled-components` package lagging behind the latest version of `styled-components`.
Note that the `// @ts-ignore` comment is temporary and should be removed once the `@types/styled-components` package is updated to match the latest version of `styled-components`.