refactor(CurrencySelect): Use React-Select styles API and react-hook-form's useController hook for CurrencySelect
The CurrencySelect component was having issues with server and client className mismatches due to the use of styled components. To fix this, the styled components were removed and the styles API of React-Select was used instead to style the component. Additionally, the useController hook provided by the react-hook-form library was used to better control the state of the CurrencySelect component and improve its reusability. The currencySelect test was also update to reflect the changes.